+++ src/main/java/com/tvm/internal/tvm_internal_project/service/onboarding/FamilyService.java
@@ -9,11 +9,11 @@ import java.util.List;
 public interface FamilyService {
     ResponseEntity<ResponseStructure<Family>> saveFamily(Family family);
 
-    ResponseEntity<ResponseStructure<Family>> getbyid(Integer id);
+    ResponseEntity<ResponseStructure<Family>> getbyid(Long id);
 
     ResponseEntity<ResponseStructure<List<Family>>> getall();
 
-    ResponseEntity<ResponseStructure<Family>> updateall(Integer id, Family family);
+    ResponseEntity<ResponseStructure<Family>> updateall(Long id, Family family);
 
-    ResponseEntity<ResponseStructure<String>> deletebyid(Integer id);
+    ResponseEntity<ResponseStructure<String>> deletebyid(Long id);
 }
